refactor: extract reducer pane geometry
